completed remedycli, added developer class, collected like terms lol
remedycli is more or less extendable, with methods like draw extrapolated to a quick_draw method where most if not all calls to draw to terminal are made. The main program loop is now listening for keys with new keys for new methods are insertable. listening for keys is now it's own method where it is now able to check at all times if the user wants to press Q to quit the program. Added developer class as possible active records example of relationships. Added a new scrape method to the scraper class to scrape a specific game's page and also set new values to the game. Added a way to navigate into a game to get more details.